refactor(frontend): Add commonSSR component
